Renamed the migration script to pulp-manage-db.
Also, this script is now responsible for loading content types to the
database instead of api.py. Added a unit test to ensure that the type
loading code gets called by this method.

I also converted the script to use exception handling instead of return
codes all over the place. There is much reformation work left to do for
this script, but this is a good start.