- 862187 /var/log/pulp/db.log now includes timestamps.
Reworked logging in the pulp-manage-db module to use the same logging
system that is used elsewhere in Pulp. This allowed us to add timestamps
to the db log with the logging configuration file, and it also allows
sysadmins to season the migration log format to taste. As part of this
commit I also removed the ability to change the log file with the former
--log-file flag, as that functionality is now made available through the
logging configuration file.