improving a test to tolerate clock rollover to the next second
The test compares two representations of a scheduled call, and each calculates
the "next_run" attribute on-the-fly. Very rarely the clock will roll over to
the next full second between those two calculations, causing them to not be
equal. This change compares them within a tolerance of 1 second instead of
requiring equality.