Fix the resource_manager for Upstart.
The resource_manager was not listening to its queue on Upstart systems.
This causes bad failure for Pulp. Fortunately, the problem was easy to
identify and fix.

Since Celery's multi worker needs to write a log file, we are unable to
eliminate /var/log/pulp/ at this time. We are, however, capturing all of
the Celery logs with Pulp's logger once it starts Pulp, so I've adjusted
the default log level on the Celery processes to ERROR from INFO. This
way, most of the log messages will only be logged through syslog.

This commit also silences Celery warnings about our use of pickle in
serializing our Task calls. We have a backlog story to convert our tasks
to use JSON serializable types, and until then it is best to silence
these warnings.