Temporarily suspend Celery logging.
This is a temporary fix to get master back into working condition. I'm
still not sure what happened, but we used to be able to capture Celery's
logs from Pulp, and now we cannot do that. I've also made sure that we
initialize Pulp at the Celery app's module level rather than waiting for
the worker_ready signal.