All server.conf settings are now optional.
Some of the settings in server.conf were not optional. It is now
possible to run Pulp with a blank server.conf with sane defaults.
Defaults are all set in pulp.server.config.

Also, pulp-dev.py used to symlink server.conf back into the developer's
workspace. Now, it copies that file into place and sets realistic
permissions on it. This way developers are free to edit settings on
their checkout without accidentially checking the changes in to the
default server.conf that gets installed by our RPM.