Adding the ability to configure SSL for Celery.
Introduced options in server.conf [tasks] that allow the necessary SSL
parameters to be specified.  Also an option added that enables SSL for the
celery_instance.

Defaults are added for these new options.  SSL for Celery is disabled by
default, but if enabled the default cert and keys are the default paths
produced by the pulp-qpid-ssl-cfg script.

A test that verifies the proper configuration of the celery_instance with SSL
and without SSL are also added.