1084716 - Register with Celery's setup_logging.
Celery provides a setup_logging signal that allows us to declare that we
will be handling the definition of the loggers and handlers. This solves
bug #1084716, wherein we were having trouble capturing all of the log
messages.

The nice thing about this commit is that we were able to remove a lot of
our code that was trying to undo some of the logging setup that Celery
had done before our start_logging() method was run. Now we can simply
define the root logger's handler.

This commit also adds a new feature where we can blacklist loggers using
pulp.server.logs.LOG_BLACKLIST. If we wanted to, we could make this a
setting in the future, but for now it's just a module level variable.