1095691 - Adding cleanup of Celery Task Results to Reaper
https://bugzilla.redhat.com/show_bug.cgi?id=1095691

The reaper is adjusted to have the document removal be done as a class method
on each model class.  To de-duplicate the vanilla behavior of that class
method, a new ReaperMixin class is introduced.  A new test is also added that
checks that each class type the reaper will encounter properly inherits from
ReaperMixin.