1093009 - Don't use symlinks for init scripts.
We had a bug[0] wherein we weren't properly starting our Celery services
upon booting. It turned out to be a combination of two problems.

One problem was that we weren't declaring a dependency on mongo, qpidd, or
rabbitmq-server in the init script, which could cause Upstart to start
our workers before it started those dependencies. Of course, if those
services are running on another machine Upstart cannot tell, so this
commit makes those dependencies optional.

The second problem was that we were using the script's name to determine
the name of the config file in /etc/default/ to use. This led to a
problem during boot because the script's name was prepended with "S50",
and there was no config file similarly prepended. The solution was to
remove what looked like a spurious line in the init script and let the
script use readlink to determine the real script name. In order to do
this, pulp_resource_manager could no longer be a symlink to
pulp_workers, so it is a first class script now that sources
pulp_workers instead. This allows us to keep the init script mostly in
pulp_workers, avoiding repeating most of the contents in there.

[0] https://bugzilla.redhat.com/show_bug.cgi?id=1093009