1092989: assume no uploads are in progress if upload working dir does not exist
Previously, pulp-admin would output an error to ~/.pulp/admin.log if "repo
uploads list" was run without a repo upload working dir. This was caused by an
uncaught OSError.

Instead, check for an exception if the repo upload working dir does not exist
(OSError w/ ENOENT), and assume no uploads are in progress if this is the case.
Other types of OSError are raised as normal.