1097247 - Add status to pulp_celerybeat script.
This commit adds a status command to the pulp_celerybeat init script.

It also fixed the pulp_workers init script to report status only on the
nodes managed by it, rather than all PID files found in /var/run/pulp/.
This also adds the ability for it to report nodes being stopped when
there are no PID files for them.

Both init scripts return 0 if the expected nodes are all running, and
1 if one or more nodes are failed or stopped.

https://bugzilla.redhat.com/show_bug.cgi?id=1097247