1102393 - Rework how we select the queue for new reservations.
This commit alters the data models for keeping track of our Celery workers. We used to track them with the model AvailableQueue, and this commit
renames that model to Worker. This commit also removes the tracking of the number of reservations by Worker, relying instead on tracking them with
the ReservedResources only. This should resolve two bugs[0][1].

[0] https://bugzilla.redhat.com/show_bug.cgi?id=1102393
[1] https://bugzilla.redhat.com/show_bug.cgi?id=1102390