1097781 - Indicate that consumer bind fails when it does.
Consumer bind tasks report success on the task even when they fail. This commit makes the CLI inspect the progress report to determine whether
binds are successful or not. There is a separate bug filed about fixing this server side.

Due to the bind task being non-standard, the CLI simply reports that the bind failed, omitting any error messages.

https://bugzilla.redhat.com/show_bug.cgi?id=1097781