1113590 - Nodes requires Pulp's cert to be trusted 1112906 - pulp-admin requires Pulp's cert to be trusted 1112904 - pulp-consumer requires Pulp's cert to be trusted
This commit addresses three bugs:

[0] - Nodes wasn't checking the signature on the Parent Pulp server
against a list of trusted certificate authorities. This commit
introduces that check, as well as a new nodes.conf pair of settings to
either disable the check or change where the certificate pack is stored.

[1] - pulp-admin also had the same issue as Nodes, and it now also has
similar settings in admin.conf.

[2] - pulp-consumer was also affected by this same security issue.
Similar settings are introduced in consumer.conf with this commit.

[0] https://bugzilla.redhat.com/show_bug.cgi?id=1113590
[1] https://bugzilla.redhat.com/show_bug.cgi?id=1112906
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1112904