1133953 - check Mongo version during startup
Previous versions of Pulp would not check the version of the Mongo database
during startup. If the version of Mongo was older than what Pulp expected,
issues would arise during runtime that could be difficult to debug.

Instead, check the Mongo version at startup and fail immediately if it's not
compatible with Pulp.

https://github.com/mongodb/mongo/blob/master/src/mongo/util/version.cpp
contains information about how Mongo versioning works.