Add a dev-setup.sh script
I created a dev-setup.sh script for testing the setup instructions from
@asmacdo. It was originally intended just to run through all his steps to test
his doc but it may be generally useful for new contributors to help them get
started quickly.

The script should work on RHEL 7+, CentOS 7+, or Fedora 19+. There is one line
that is hard-coded for Fedora 20 but it can go away once we push Pulp 2.5.1
alpha into the testing repo.

This commit also has a couple of step order fixes to the document.