721314 - add man pages for pulp-admin and pulp-consumer
Docs for pulp-admin and pulp-consumer already exist but were only being
rendered as HTML and not man pages.

This commit adds a build step to create man pages and also adds the man pages
to pulp-admin and pulp-consumer.

Note that intersphinx will fail in build environments with limited network
connectivity. It should fail with a warning though and not an error.