Updates to /status API
This patch returns additional data from the /status API call. Specifically, it
returns a worker list, the message broker connection status, the database
connection status, and the Pulp version (RHBZ #704184).

There is a small behavior change to include the resource manager and scheduler
as part of the worker heartbeat list. Previously, these were not included in
the list. These entries are filtered out when Pulp requests a free worker.

A few PEP8 fixes are included as well for files that were touched in this patch.

Additionally, the `ping()` method in ServerInfoAPI has been removed; this
called a URL that no longer exists.