Adds local working directories for pulp workers
This patch adds ability to specify 'working_directory' setting in the
/etc/pulp/server.conf [server] section. The default value is
/var/cache/pulp. This patch also modifies the pulp-celery SELinux policy.
The spec file is also modified to create the default working directory when
the pulp-server rpm is installed. Pulp workers can now request a working
directory when performing tasks.  Each directory is removed after a task
enters a final state.