Move repo auth functionality to platform
This code previously existed only in the RPM plugin but other plugins need to
use it as well.

This patch does the following:

 * moves code and tests from pulp_rpm.repo_auth to pulp.repoauth
 * moves the repo_auth wsgi script and conf file to platform
 * updates the pulp spec file
 * adds a new package, python-pulp-repoauth
 * updates pulp-dev.py
 * updates to the playpen scripts to create test certs
 * adds unit tests to improve coverage of repoauth module
 * adds an optional entry to repo_auth.conf to handle multiple URL prefixes
 * release notes for 2.7.x (assuming this goes into a Y release)

This patch also bumps the testing cert expiration dates.

The old certs expired on Jan 24 2015. This patch replaces the test certificates
with a new set that expires in 2018.

To regen, use `create_chain_data.sh`