540 - Check previous version of pulp-server SELinux policy at ugrade
This patch enhances the pulp-selinux RPM install process. The version of the
pulp-server SELinux policy is recorded in /var/lib/rpm-state/pulp before the new
policy is installed. Then that version is passed as an argument to relabel.sh.
The relabel script then only relabels the files that have not been properly labeled
in the previously installed version. If no version is passed in, all files
are relabeled.

Fixes #540