convert "workers" collection to mongoengine
This patch converts the `workers` collection to mongoengine. It also moves some
functionality into tasks.py that previously lived in the worker manager.

One thing to note, it appears that the existing workers collection did not have
a `_ns` field. Thus, I did not add this into the new collection.

This commit includes a fix to CriteriaQuerySet's `update()` method which was
incorrectly firing the `post_save()` hook on TaskStatus objects, even when the
Document was not a task status.

A spec file requirement for `python-blinker` was also added to support
mongoengine signals. Additionally, a small unrelated flake8 fixup to
test_users.py was added.

fixes #88