More robust hostname default, prompt for input
Remote hosts are more likely to be connecting to qpidd using the fully
qualified hostname, especially in complex environments where not all
machines share the same domain/search suffix.

Sometimes the fqdn is not properly set up, and the hostname contains the
entire fqdn. This change addresses this, while also allowing the user to
specify in case a DNS alias or IP address is preferred.

It also fills the example configs out with the hostname used as the CN
in the cert, as that is the only string that will actually work due to
server certificate verification.

Support relative paths for INST_DIR

Previously relative path usage was not supported for INST_DIR due to the
script cding to the temp directory. Use readlink to make it an absolute
path.