Allow passwordless mongodb connections to be configured
MongoDB allows passwordless logins. However, Pulp was checking that both the
username *and* password were set before attempting a login.

This patch alters the user/pass check to fail only if a password is set sans
username. If a username is set sans password, a MongoDB connection will be
attempted.

Note that all of the auth still happens in Mongo. This was just a "pre-check"
that Pulp was doing to give a clear error message.

fixes #708