Allow setting the broker_login_method.
When using RabbitMQ with SSL client certificates, it is necessary to use
the rabbitmq_auth_mechanism_ssl which uses the common_name (CN) of the
client certificate as user name and the use SASL EXTERNAL to have
RabbitMQ pick up the configuration.

Unfortunately, by default, pulp only supports the default (AMQPLAIN)
authentication and it can not be configured.

This change adds a new configuration parameter to the tasks section
which allows setting the requested authentication mechanism through the
celery BROKER_LOGIN_METHOD configuration setting.

This allows using RabbitMQ with strict SSL client certificates:

[
  {rabbit, [
    {ssl_listeners, [5671]},
    {auth_mechanisms, ['EXTERNAL']},
    {ssl_options, [
      {cacertfile,           '.../cacertfile'},
      {certfile,             '.../cert.crt'},
      {keyfile,              '.../cert.key'},
      {verify,               verify_peer},
      {fail_if_no_peer_cert, true}
    ]},
    {ssl_cert_login_from,  common_name}]
  }
].

and adding the CN values from the client certificates as users to RabbitMQ.