Adjusts the upload delete API implementation
The upload delete method used inode data to determine if a file
exists before deleting it. On filesystems with cached or slow-to-
respond inode data like NFS this was causing issues. The new
implementation always tries to remove a file and silences
only the error in the case where the file does not exist.

Also some small flake8 fixes.

closes #480

https://pulp.plan.io/issues/480