Update docs for task notification messages
The previous docs were not suitable for someone who wanted to integrate with
Pulp but was not already familiar with TaskStatus objects.

This patch adds some examples and further explanation on task notifications and
how to use them. It also adds some comments to server.conf about the new
options.

re #87