Ensure that tests share a single DB conenction.
Many of our tests independently connected to the database, and they did
not always ensure that they were connecting to the test database. This
commit fixes the tests so they all use the test database.

Additionally, this commit refactors all of the entry points such that
the connection is no longer started as a side effect of importing a
module. Most of our code already worked this way, but this commit
refactors the Celery workers to start the database connection using a
signal instead of through an import.

This commit also makes it an error for the connection to be initialized
more than once.

https://pulp.plan.io/issues/131
https://pulp.plan.io/issues/460

fixes #131
fixes #460