Create the test config in code rather than a file.
This commit sets the test config in the test setup code, rather than
importing the configuration from a test file. This is easier to edit,
and also does not require the pulp.devel package to distribute a data
file.

Furthermore, this commit also forbids tests to call
pulp.server.config.config.set() or any variants thereof. Instead, it
provides a new convenience mock function that makes it easy to patch
particular settings during tests without altering the settings globally
as was done before.

Many tests had to be fixed as a result of these changes, but now they
are not altering global settings as they were doing before.

https://pulp.plan.io/issues/940

fixes #940