better support for installing standalone repoauth
Previously, the pulp repoauth RPM was not installable without also installing
the rest of pulp-server. This was not ideal for instances that served protected
content to client systems, but did not run full Pulp installations.

This patch does the following:

 * moves an openssl lib from `pulp.server.common` to `pulp.repoauth`. Note that
   we couldn't move it to `pulp.common` since our openssl code has some calls to
   `subprocess` that do not work on Python 2.4.

 * adds a few deps to python-pulp-repoauth that previously were provided by
   pulp-server's deps

 * fixes an unrelated issue where a "real" conf file was used in a few unit
   tests

I confirmed that no plugins are using `pulp.server.common.*`.

Fixes #975

https://pulp.plan.io/issues/975