Update our troubleshooting for qpid-tools.
Qpid has unfortunately moved the qpidtoollibs Python package from the
python-qpid-qmf package to the qpid-tools package. This commit adds the
qpid-tools package to the troubleshooting guide in addition to the
python-qpid-qmf. We need to mention both for now since we do not know if
our users will be using the newer or older version of qpid.