Remove the -e from the bash provisioner.
For some reason, the -e started causing provisioning failures recently.
I suspect that a Fedora update might have changed the behavior of that
flag, but I have not confirmed this. Since we are slowly switching from
the shell provisioner to Ansible I believe it is best to disable this
flag for now to get around the immediate issue.