Convert more of the dev provisioning to Ansible.
This commit converts more of the Vagrant provisioning from bash to
Ansible. It also removes the hard dependency on pulp_rpm in the
remaining bash provisioning. Finally, it updates the Vagrant
documentation to be more current to reflect the change on the dependency
on pulp_rpm, as well as correcting some notes that are no longer
accurate.