Properly handles ['database']['seeds'] config
This patch enables pulp to iterate over a list of seeds until a connection
is made. When more than one seed is present, an exception is raised if no
'replica_set' config is specified. An exception is also raised if an empty
string is passed in as value for 'seeds'.

re #1012
https://pulp.plan.io/issues/1012