fixed orphan removal by type when the type uses mongoengine
Several places in the code, only two of which I touched here, need to know the
unit key fields for a given type ID. I created a new controller function that
takes care of finding that information regardless of whether the type is
defined the new way (mongonengine) or old way. The next step will be to use
that controller function in additional places, but that is beyond the scope of
this bug fix.

fixes #1215