ContentUnit now models pulp_user_metadata as pulp_user_metadata
The first-pass implementation of ContentUnit for mongoengine
used an attribute name of 'user_metadata' to model the database
field named 'pulp_user_metadata'. This is a nice chance, but it
forces multiple plugins to handle name translation for both
Criteria objects and serializers.

By modeling the attribute as the actual database name everything
becomes simpler. We can rename pulp_user_metadata to
user_metadata in a future X release.