Renames last_updated to _last_updated on ContentUnit
The first-pass implementation of ContentUnit for mongoengine
used an attribute name of 'last_updated' to model the database
field named '_last_updated'. This is a nice chance, but it
forces multiple plugins to handle name translation for both
Criteria objects and serializers.

By modeling the attribute as the actual database name everything
becomes simpler. We can rename _last_updated to last_updated in
a future X release.