Add Celery flower to the Vagrant dev box.
Flower is a web-based tool for monitoring and administrating Celery
clusters. This adds Flower to the Ansible dev playbook and includes a
systemd unit for the service. It is configured to run on port 5555.

This also adds hostname management to the example Vagrantfile. If this
is enabled, it is possible to access the Flower dashboard from your host
machine at http://dev.example.com:5555/.