Fix the type of Repo model expected in the publish steps.
The publish steps were inconsistent in which of the two Repository models they worked with, mixing the APIs.
This commit documents that they only work with pulp.plugins.model.Repository and fixes an API usage so that
it works with that model instead of with the pulp.server.db.model.Repository model.