The lazy catalog is purged during an orphan purge.
This change appears in both the mongoegine version and the
non-mongoengine version of the "delete orphan by type" method because
although lazy is only supported by mongoengine types, the
``delete_orphan_by_id`` method works with mongoengine types and calls
the non-mongoegine version of "delete orphan by type".

It also updates the mongoengine version of "delete by type" to match the
signature of the non-mongoengine version.

closes #1339