Parallelizes applicability regeneration for updated repository
This patch provides a new Celery task for performing applicability regenration
for a batch of applicability profiles. The ApplicabilityRegenerationManager
dispatches a series of tasks with the same group id. Each task is dispatched with
a list of up to 10 RepoProfileApplicabilities to reevaluate.

The API endpoint for generating content applicability for updated repositories
changes as part of this patch. Instead of returning 202 with a call report, the
server returns 202 with a group call report.

This patch does not make any changes to the algorithm used to calculate content
applicability.

https://pulp.plan.io/issues/20
closes #20