Allow developers to override the GetLocalUnitsStep's available_units.
Some types are a mixture of Units and Metadata, and need to be treated separately. The GetLocalUnitsStep did
not allow developers to specify what data structure would contain the iterable of available Units. This
commit adds the ability for developers to give a pointer to that data structure while maintaining the
previous behavior if it is not specified for backwards compatibility.

This is needed for the Docker plugin Mongoengine conversion as it reduces the number of steps that are needed
in its importer's code.

https://pulp.plan.io/issues/863

re #863