drop checks for dropDups
The upgrade to fedora 23 revealed that the 'dropDups' property of mongo
indices was deprecated in 2.6 and removed in 2.7.

The removed 'dropDups' assertions weren't stricly necessary, since the
test is checking that the uniqueness contraint is applied, which it is.

dropDups only comes into play if it's explicitly requested, which we
never do.

https://pulp.plan.io/issues/1393

fixes #1393