Use Celery's local_received time for worker availability.
Pulp had been using the timestamp attribute of Celery's Events to determine when the workers were most
recently available. Due to a bug in Celery, this cause an off-by-one-hour bug for users who live in a locale
that does daylight savings time that is not currently observing daylight savings time (such as New York in
the winter):

https://github.com/celery/celery/issues/1802#issuecomment-161916587

This fixes the issue by using the local_received time which is computed in the receiver by Celery with
time.time() without any timezone shifting attempts.

https://pulp.plan.io/issues/1380

fixes #1380