Introducing from_dict() and to_dict()
closes #23
https://pulp.plan.io/issues/23

UnitAssociationCriteria arrives as dict in the task code, when it needs to be
UnitAssociationCriteria object. to_dict() method was introduced to serialize
UnitAssociationCriteria to a dictionary prior to it being sent as an argument
to apply_sync(). the from_dict() method was introduced to return UnitAssociationCriteria
object from a dictionary that was produced by to_dict()

This commit also introduces a testcase that checks that all Celery tasks defined in Pulp
inhherit from PulpTask and not from Celery directly.

There have been done also changes to criteria. Due to the introduction of de/sereliazions
methods, criteria's value cannot be None. Code and tests were updated accordigly.