Ansible debug tasks for dev role, disable cachier
bmbouter and I were doing some debugging with gdb, and it was a little
annoying to have to install all the debug tools I wanted for the job.
There are a *ton* of packages, though, so I made the tasks to install
debug tools conditional, and added a comment to the example Vagrantfile
that can enable them.

I also put in my suggested short-term fix of commenting out the cachier
conditional while we wait for that issue to be resolved.