Update the HTTP callback event publish to use requests.
This addresses two problems: firstly, the listener would
not validate the signature on the server's x509 certificate. Secondly,
it was willing to perform SSLv2 and SSLv3. python-requests validates
certificates by default. On EL6+ and Fedora, it uses the system trust
store to perform this validation.

The protocols supported by requests is determined by urllib3. As of
Fedora 23, SSLv2 and SSLv3 is not supported.

closes #581