adding a call to rebuild unit counts at the end of every associate_from_repo call
There was a bug in at least one plugin (and probably more) that forgot to
account for the unit counts during mongoengine conversion. This appears to be a
relatively inexpensive call, and I think is a fine and robust approach instead
of the previous one, where pulp tracked unit count deltas for each operation.