Sends back a group call report for summary of task group with 0 tasks
Originally we wanted to return  a 404 when querying for a task group that has no tasks
associated with it. However, we always return a task group id when kicking off the
applicability regeneration task. User should be able to see that 0 tasks were dispatched
as part of that task group.

https://pulp.plan.io/issues/1582
closes #1582