Remove database initialization from content.wsgi.
Firstly, the content application does not use the database, so it
shouldn't be initializing a connection to the database. Secondly, this
implementation breaks with pymongo-3.x because it was creating a
connection in a parent process (pre-WSGI forking, I believe) which is
not safe (see https://jira.mongodb.org/browse/PYTHON-961).