Rework the vagrant-setup.sh script to only install once.
The vagrant-setup.sh script had been installing, uninstalling,
and reinstalling the plugins which took a long time. This was done
for the DRY principle so that some code could be shared with preset
and vagrant-setup.sh, but it cost a few more minutes on each
vagrant up. This commit reworks preset() so that the more complex
operations are split into a common function that vagrant-setup.sh
can still use, which allows it not to call preset.