Create a simple directory index for the content.wsgi app.
In the past we were able to use Apache's directory index feature.
However, for lazy content we pass the request through to the content
WSGI application and are unable to use this feature. This introduces a
simple template that lists the directories and files within a requested
directory, as well as clear HTTP 404 and HTTP 500 error pages.

closes #1512