Set Apache to allow 10,000 requests per TCP connection.
By default, Apache allows 100 HTTP requests per TCP connection before
resetting it. Since it is not uncommon for Pulp to serve thousands or
tens of thousands of requests from a single client, it makes sense to
not use the default. This is especially noticeable when performing
Pulp-to-Pulp syncs.

This is related to issue #1629.