Use SSL_CLIENT_CERT and fall back to mod_ssl.var_lookup.
Depending on how a request arrives and who is performing the
authentication, `mod_ssl.var_lookup` may not be present and
`SSL_CLIENT_CERT` is. This commit causes the authenticator to look in
both locations for the certificate and if neither can be found, it
returns `False`. Previously, it would cause an HTTP 500 by raising a
KeyError.

closes #1667