Refine playpen/all_27.sh script
The `all_27.sh` script walks through each repository in `~/devel/pulp*`, checks
out an appropriate branch, and either installs or uninstalls the code in that
branch using the `pulp-dev.py` script. This works so long as all of the
repositories found are, in fact, either Pulp or one of its plugins. pulp-smash
is neither. Refine the script to make it more precisely find correct
repositories.