Run psmash with python3 and py.test.
The development environment used to run psmash with Python 2. In
order to more easily catch Python 3 incompatibilties, this commit
adjusts the psmash command to create the pulp-smash virtualenv with
Python 3 instead. Unfortunately, due to pulp-smash #146[0],
it does not pass when run with unittest2. Thus, this commit also
switches psmash to use py.test as a workaround.

[0] https://github.com/PulpQE/pulp-smash/issues/146