Modify the content web view to account for symbolic links
Users sometimes make /var/lib/pulp a symbolic link. This worked fine
until we introduced the content wsgi application for lazy syncing. This
modifies it to account for symbolic links when deciding whether or not
it's safe to serve a file. It also documents the necesssary Apache
changes that need to take place to make symbolic links work.

closes #1791