Safely create tmp dir for the Nodes certificate (CVE-2016-3108).
Security researcher Sander Bos contacted the Pulp team to notify us
that the pulp-gen-nodes-certificate script suffers from the same
exploit as was found in CVE-2016-3095, namely that the $TMP
directory that contains the Nodes private key was created in an
unsafe manner. This commit contains his proposed patch, using
mktemp -d to safely create the directory.

Additionally, I added a set -e so that the script would exit upon
error.

Thanks to Sander Bos for taking the time to carefully inspect the
Pulp codebase and for writing a wonderfully detailed report
describing the issue and the fix for it.

Credit also goes to Jeremy Cline (Red Hat) for independently
reporting this issue.

https://pulp.plan.io/issues/1830

fixes #1830