pulp.spec now generate RSA keys with umask 077 (CVE-2016-3111)
During installation, the RSA key pairs used to validate messages between
the pulp server and pulp consumers were generated in a directory that is
world-readable with a umask of 002. After it was written, the permissions
were modified to protect the key. For a brief moment, the RSA keys were
world-readable. This commit explicitly sets the umask in the %post
scriptlet to be 077 so it is only readable to the owner.

https://pulp.plan.io/issues/1837

fixes #1837